refactor(structure): sort all file types into their corresponding folders
