+++ src/features/audience/api/jobStorage.ts
@@ -1,4 +1,4 @@
-import type { AudienceType } from "./api.type";
+import type { AudienceType } from "@/features/audience/model/type";
 
 function storageKey(login: string, type: AudienceType): string {
 	return `audience-job:${login}:${type}`;
