+++ src/features/user-profile/model/type.ts
@@ -1,0 +1,8 @@
+export interface UserProfileResponse {
+	login: string;
+	name: string | null;
+	avatarUrl: string;
+	url: string;
+	followersCount: number;
+	followingCount: number;
+}
